LayoutTests/svg/custom/use-event-listener-once-option-preserved.html

<!DOCTYPE html>
<html>
<body>
<svg id="svg" width="200" height="200">
    <defs>
        <rect id="target" width="200" height="200" fill="green"/>
    </defs>
</svg>
<script src="../../resources/js-test.js"></script>
<script>
description("Tests that the 'once' option on an event listener is preserved when copied to an SVG use shadow tree.");

var count = 0;

// Add a { once: true } listener to the original element FIRST.
var target = document.getElementById("target");
target.addEventListener("test", function() { count++; }, { once: true });

// Now dynamically create the <use> element so the shadow tree is built
// AFTER the listener exists, ensuring transferEventListenersToShadowTree
// copies it to the shadow clone.
var svg = document.getElementById("svg");
var use = document.createElementNS("http://www.w3.org/2000/svg", "use");
use.setAttribute("href", "#target");
svg.appendChild(use);

// Force layout to build the shadow tree (which copies listeners).
document.body.offsetHeight;

// Access the shadow clone via internals and dispatch events to it.
var shadowRoot = internals.shadowRoot(use);
var shadowClone = shadowRoot.querySelector("rect");
shadowClone.dispatchEvent(new Event("test"));
shadowClone.dispatchEvent(new Event("test"));

// With once properly preserved, the listener should fire only once.
shouldBe("count", "1");
</script>
</body>
</html>
