LayoutTests/svg/custom/use-event-listener-once-option-preserved-expected.txt

Tests that the 'once' option on an event listener is preserved when copied to an SVG use shadow tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS count is 1
PASS successfullyParsed is true

TEST COMPLETE
